pyproject.toml
[project]
name = "srcnew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audiomentations>=0.35.0",
    "dotenv>=0.9.9",
    "icecream>=2.1.4",
    "librosa>=0.10.2.post1",
    "matplotlib>=3.10.1",
    "peft>=0.14.0",
    "pip>=25.0.1",
    "pydantic>=2.10.6",
    "pytorch-lightning>=2.5.0.post0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "telebot>=0.0.5",
    "torchaudio>=2.6.0",
    "torchinfo>=1.8.0",
    "torchvision>=0.21.0",
    "torchviz>=0.0.3",
    "transformers>=4.49.0",
    "wandb>=0.19.7",
]
